add class and methods

+++ src/main/java/com/example/petproject/dao/ExecutorRepository.java
@@ -1,9 +1,13 @@
 package com.example.petproject.dao;
 
 import com.example.petproject.model.Executor;
+import com.example.petproject.model.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface ExecutorRepository extends JpaRepository<Executor, Integer> {
+public interface ExecutorRepository extends JpaRepository<Executor, Long> {
+//    List<Executor> findByPersons(List<Person> persons);
 }
